requirements.txt
Flask
python-dotenv
langchain-community
